requirements.txt
Flask==2.2.5
Flask-PyMongo==2.3.0
dnspython==2.3.0
python-dotenv==1.0.0
pillow==12.0.0
Flask-Login==0.6.2
werkzeug==2.2.3